+++ CHANGELOG.md
@@ -14,8 +14,9 @@ Updated/Enhanced Rules and Bug Fixes
  - #273: **UnnecessaryToString** rule: Add check for `xx.toString()` added (+) to a String constant; add check for GString expression `${xxx.toString()}`.
 
 CodeNarc Plugin Mechanism
+ - #494: Add *plugins* property to the CodeNarc Ant Task (`CodeNarcTask`) and `-plugins` parameter to the command-line.
+ - #494: `CodeNarcRunner`: Apply plugins for rules, violations and reports. Register any plugin classes specified by the "org.codenarc.plugins" system property.
  - #494: Add `sourceCode` property to `FileResults`; set it from `FilesystemSourceAnalyzer` and `AntFileSetSourceAnalyzer`.
- - #494: CodeNarcRunner: Apply plugins for rules, violations and reports.
 
 Framework and Infrastructure
  - #525: Upgrade to Groovy 2.5.12. [BREAKING CHANGE]